// part9/flight-diary-frontend/src/types.ts
export interface HeaderProps {
	title: string;
}

export enum Visibility {
	Great = "great",
	Good = "good",
	Ok = "ok",
	Poor = "poor",
}

export enum Weather {
	Sunny = "sunny",
	Rainy = "rainy",
	Cloudy = "cloudy",
	Stormy = "stormy",
	Windy = "windy",
}

export interface DiaryEntry {
	id?: number;
	date: string;
	weather: Weather;
	visibility: Visibility;
	comment: string;
}

export type NonSensitiveDiaryEntry = Omit<DiaryEntry, "comment">;
export type NewDiaryEntry = Omit<DiaryEntry, "id">;

export interface ContentProps {
	entries: NonSensitiveDiaryEntry[];
}